+++ e2e/project/frontity.settings.ts
@@ -67,7 +67,7 @@ const settings: Settings = [
       },
       {
         name: "@frontity/wp-source",
-        state: { source: { api: "https://test.frontity.org/wp-json" } },
+        state: { source: { url: "https://test.frontity.org/" } },
       },
     ],
   },
@@ -136,7 +136,7 @@ const settings: Settings = [
       "@frontity/wp-comments",
       {
         name: "@frontity/wp-source",
-        state: { source: { api: "http://localhost:8080/wp-json" } },
+        state: { source: { url: "http://localhost:8080/" } },
       },
     ],
   },
@@ -198,7 +198,52 @@ const settings: Settings = [
         name: "@frontity/wp-source",
         state: {
           source: {
-            api: "http://localhost:8080/wp-json",
+            url: "http://localhost:8080/",
+          },
+        },
+      },
+    ],
+  },
+  {
+    name: "smart-adserver",
+    packages: [
+      "e2e-smart-adserver",
+      "@frontity/tiny-router",
+      {
+        name: "@frontity/smart-adserver",
+        state: {
+          smartAdserver: {
+            networkId: 1445,
+            subdomain: "www",
+          },
+          fills: {
+            smartAdserver: {
+              headerAd: {
+                slot: "header",
+                library: "smartAdserver.SmartAd",
+                priority: 5,
+                props: {
+                  callType: "std",
+                  siteId: 78061,
+                  pageId: 884496,
+                  formatId: 33780,
+                  tagId: "hello",
+                },
+              },
+              bottomAd: {
+                slot: "bottom",
+                library: "smartAdserver.SmartAd",
+                priority: 5,
+                props: {
+                  callType: "std",
+                  siteId: 78061,
+                  pageId: 884496,
+                  formatId: 33780,
+                  minHeight: 100,
+                  tagId: "std-min-height",
+                },
+              },
+            },
           },
         },
       },
@@ -218,7 +263,7 @@ const settings: Settings = [
         name: "@frontity/wp-source",
         state: {
           source: {
-            api: "http://localhost:8080/wp-json",
+            url: "http://localhost:8080/",
             postTypes: [
               {
                 type: "movie",
@@ -260,7 +305,7 @@ const settings: Settings = [
         name: "@frontity/wp-source",
         state: {
           source: {
-            api: "http://localhost:8080/wp-json",
+            url: "http://localhost:8080/",
             postTypes: [
               {
                 type: "movie",
@@ -288,6 +333,31 @@ const settings: Settings = [
             transformLinks: {
               base: "http://localhost:8080",
             },
+          },
+        },
+      },
+    ],
+  },
+  {
+    name: "embedded-mode",
+    match: [process.env.FRONTITY_SERVER],
+    state: {
+      frontity: {
+        url: "http://localhost:8080",
+      },
+    },
+    packages: [
+      "e2e-preview",
+      {
+        name: "@frontity/wp-source",
+        state: {
+          source: {
+            postTypes: [
+              {
+                type: "movie",
+                endpoint: "movies",
+              },
+            ],
           },
         },
       },