+++ packages/wp-source/src/libraries/handlers/postType.ts
@@ -1,6 +1,7 @@
 import { Handler } from "../../../types";
 import capitalize from "./utils/capitalize";
 import { ServerError } from "@frontity/source";
+import { PostTypeData } from "@frontity/source/types/data";
 
 /**
  * The parameters for {@link postTypeHandler}.
@@ -47,9 +48,14 @@ const postTypeHandler = ({
   libraries,
   force,
 }) => {
+  // Name of the endpoint that returned an entity.
+  // Used later in case this fetch is for a preview.
+  let matchedEndpoint = "";
+
   // 1. search id in state or get the entity from WP REST API
   const { route, query } = libraries.source.parse(link);
-  if (!state.source.get(route).id || force) {
+  const routeData: Partial<PostTypeData> = state.source.get(route);
+  if (!routeData.id || force) {
     const { slug } = params;
 
     // 1.1 transform "posts" endpoint to state.source.postEndpoint
@@ -79,6 +85,7 @@ const postTypeHandler = ({
         if (populated[0].link === route) {
           isHandled = true;
           isMismatched = false;
+          matchedEndpoint = endpoint;
           break;
         } else {
           isMismatched = true;
@@ -102,7 +109,7 @@ const postTypeHandler = ({
   }
 
   // 2. get `type` and `id` from route data and assign props to data
-  const { type, id } = state.source.get(route);
+  const { type, id }: Partial<PostTypeData> = state.source.get(route);
   const data = state.source.get(link);
   Object.assign(data, {
     type,
@@ -111,7 +118,38 @@ const postTypeHandler = ({
     id,
     isPostType: true,
     [`is${capitalize(type)}`]: true,
-  });
+  }) as PostTypeData; // This ensures the resulting type is correct.
+
+  // Overwrite properties if the request is a preview.
+  if (query.preview && state.source.auth) {
+    // Get entity from the state.
+    const entity = state.source[type][id];
+
+    // Fetch the latest revision using the token.
+    const response = await libraries.source.api.get({
+      endpoint: `${matchedEndpoint}/${id}/revisions?per_page=1`,
+      params: state.source.params,
+      auth: state.source.auth,
+    });
+
+    // Get modified props from revision.
+    const revision = await response.json();
+    if (revision.code) {
+      console.log(revision);
+      throw new ServerError(revision.message, revision.data.status);
+    }
+
+    const [json] = revision;
+
+    if (json.parent === id) {
+      const { title, content, excerpt } = json;
+      // Merge props with entity.
+      Object.assign(entity, { title, content, excerpt });
+    } else {
+      // Error response.
+      console.warn(json);
+    }
+  }
 };
 
 export default postTypeHandler;